Changed AmazonPurchase, Labor, and Quote

+++ Release_Candidate_Ervin_Hostetler/Models/AmazonPurchase.cs
@@ -10,8 +10,8 @@ namespace Release_Candidate_Ervin_Hostetler.Models
         public int AmazonPurchaseId { get; set; }
         public int Quantity { get; set; }
         public string Description { get; set; }
-        public int PurchasePrice { get; set; }
+        public double PurchasePrice { get; set; }
         public string URL { get; set; }
-        public int ShippingCost { get; set; }
+        public double ShippingCost { get; set; }
     }
 }
